+++ python/pathway/tests/test_common.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import functools
 import inspect
-import multiprocessing
 import os
 import pathlib
 import re
@@ -30,6 +29,7 @@ from pathway.tests.utils import (
     assert_table_equality_wo_index,
     assert_table_equality_wo_index_types,
     assert_table_equality_wo_types,
+    mp_fork,
     needs_multiprocessing_fork,
     run_all,
     warns_here,
@@ -6524,7 +6524,7 @@ def test_remove_retractions():
 
 @needs_multiprocessing_fork
 def test_termination_after_fork():
-    p = multiprocessing.Process(target=pw.run)
+    p = mp_fork.Process(target=pw.run)
     p.start()
     try:
         p.join(timeout=1)